Set buffer log to debug level
The buffer exceeded log was reporting at the error level. Once
the conditions are met, this log can pollute the logs. The
pause/resume mechanism was implemented to handle when the
buffer is full and it should not require user intervention.
This log is being moved to debug level to reduce noise and
confusion.