Add deadline to back pressure wait
The connection writer calls Wait before writing the connection. If the
back pressure was paused due to the read buffer exceeding the buffer
limit, it will call Wait on the sync Cond. However, if the connection
was dropped but not explicitly closed, the goroutine will hang and never
wake up again.

This patch works around the issue by adding a deadline to wake up the
condition. If the deadline is exceeded, the connection is formally
closed.